changing resources for multi language courses

+++ app/Http/Controllers/Admin/AdminCourseController.php
@@ -8,6 +8,7 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\File;
 use App\Http\Requests\CourseCreateRequest;
+use App\Models\EnglishCourse;
 
 class AdminCourseController extends Controller
 {
@@ -21,35 +22,40 @@ class AdminCourseController extends Controller
     public function all()
     {
         $courses = Course::all();
-        return view("admin.courses.all", compact("courses"));
+        $englishcourses = EnglishCourse::all();
+        return view("admin.courses.all", compact("courses", "englishcourses"));
     }
 
     //  for showing online courses
     public function online()
     {
         $courses = Course::where("mode", 1)->get();
-        return view("admin.courses.online", compact("courses"));
+        $englishcourses = EnglishCourse::where("mode", 1)->get();
+        return view("admin.courses.online", compact("courses", "englishcourses"));
     }
 
     //  for showing offline courses
     public function offline()
     {
         $courses = Course::where("mode", 0)->get();
-        return view("admin.courses.offline", compact("courses"));
+        $englishcourses = EnglishCourse::where("mode", 0)->get();
+        return view("admin.courses.offline", compact("courses", "englishcourses"));
     }
 
     //  for showing free courses
     public function free()
     {
         $courses = Course::where("price", 0)->get();
-        return view("admin.courses.free", compact("courses"));
+        $englishcourses = EnglishCourse::where("price", 0)->get();
+        return view("admin.courses.free", compact("courses", "englishcourses"));
     }
 
     //  for showing not_free courses
     public function notfree()
     {
         $courses = Course::where("price", ">", 0)->get();
-        return view("admin.courses.notfree", compact("courses"));
+        $englishcourses = EnglishCourse::where("price", ">", 0)->get();
+        return view("admin.courses.notfree", compact("courses", "englishcourses"));
     }
 
 
@@ -81,38 +87,77 @@ class AdminCourseController extends Controller
      */
     public function store(Request $request)
     {
-        $course = new Course();
-        $course->title = $request->title;
-        $course->price = $request->price;
-        $course->master_name = $request->master_name;
-        $course->master_job = $request->master_job;
-        if ($request->link !== null) {
-            $course->link = $request->link;
-        }
-        $course->introduction_v_link = $request->introduction_v_link;
-        if ($request->off > 0) {
-            $course->off = $request->off;
-        }
-        $course->type = $request->type;
-        $course->mode = $request->mode;
-        $course->introduction = $request->introduction;
-        $course->description = $request->description;
-        $course->licensable = $request->licensable;
-        $course->meta_key_words = $request->meta_key_words;
-        $course->meta_descriptions = $request->meta_descriptions;
-        $course->save();
+        // if its persian
+        if ($request->lang == 0) {
+            $course = new Course();
+            $course->title = $request->title;
+            $course->price = $request->price;
+            $course->master_name = $request->master_name;
+            $course->master_job = $request->master_job;
+            if ($request->link !== null) {
+                $course->link = $request->link;
+            }
+            $course->introduction_v_link = $request->introduction_v_link;
+            if ($request->off > 0) {
+                $course->off = $request->off;
+            }
+            $course->type = $request->type;
+            $course->mode = $request->mode;
+            $course->introduction = $request->introduction;
+            $course->description = $request->description;
+            $course->licensable = $request->licensable;
+            $course->meta_key_words = $request->meta_key_words;
+            $course->meta_descriptions = $request->meta_descriptions;
+            $course->save();
 
-        // saving image in image table
-        $image = new Image();
-        $imagename = time() . "." . $request->image->extension();
-        $filename = $course->title . "." . $course->id;
-        $request->image->move(public_path("photos/courses/$filename/"), $imagename);
-        $image->name = $request->image_name;
-        $image->alt = $request->alt;
-        $image->uploader_id = auth()->user()->id;
-        $image->path = "photos/courses/$filename/$imagename";
-        $course->images()->save($image);
-        // saving image in image table
+            // saving image in image table
+            $image = new Image();
+            $imagename = time() . "." . $request->image->extension();
+            $filename = $course->title . "." . $course->id;
+            $request->image->move(public_path("photos/courses/$filename/"), $imagename);
+            $image->name = $request->image_name;
+            $image->alt = $request->alt;
+            $image->uploader_id = auth()->user()->id;
+            $image->path = "photos/courses/$filename/$imagename";
+            $course->images()->save($image);
+            // saving image in image table
+        }
+
+        // if its english
+        if ($request->lang == 1) {
+            $course = new EnglishCourse();
+            $course->title = $request->title;
+            $course->price = $request->price;
+            $course->master_name = $request->master_name;
+            $course->master_job = $request->master_job;
+            if ($request->link !== null) {
+                $course->link = $request->link;
+            }
+            $course->introduction_v_link = $request->introduction_v_link;
+            if ($request->off > 0) {
+                $course->off = $request->off;
+            }
+            $course->type = $request->type;
+            $course->mode = $request->mode;
+            $course->introduction = $request->introduction;
+            $course->description = $request->description;
+            $course->licensable = $request->licensable;
+            $course->meta_key_words = $request->meta_key_words;
+            $course->meta_descriptions = $request->meta_descriptions;
+            $course->save();
+
+            // saving image in image table
+            $image = new Image();
+            $imagename = time() . "." . $request->image->extension();
+            $filename = $course->title . "." . $course->id;
+            $request->image->move(public_path("photos/courses/$filename/"), $imagename);
+            $image->name = $request->image_name;
+            $image->alt = $request->alt;
+            $image->uploader_id = auth()->user()->id;
+            $image->path = "photos/courses/$filename/$imagename";
+            $course->images()->save($image);
+            // saving image in image table
+        }
         return redirect()->route("admin.courses.all")->with("success", ".دوره شما با موفقیت ساخته شد");
     }
 
@@ -133,10 +178,19 @@ class AdminCourseController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit($id, $lang)
     {
-        $course = Course::find($id);
-        return view("admin.courses.edit", compact("course"));
+        // check if course is farsi
+        if ($lang == 0) {
+            $course = Course::find($id);
+            return view("admin.courses.edit", compact("course", "lang"));
+        }
+
+        // check if course is english
+        if ($lang == 1) {
+            $course = EnglishCourse::find($id);
+            return view("admin.courses.edit", compact("course", "lang"));
+        }
     }
 
     /**
@@ -148,27 +202,57 @@ class AdminCourseController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $course = Course::find($id);
-        $course->title = $request->title;
-        $course->price = $request->price;
-        $course->master_name = $request->master_name;
-        $course->master_job = $request->master_job;
-        if ($request->link !== null) {
-            $course->link = $request->link;
+
+        // check if course is farsi
+        if ($request->lang == 0) {
+            $course = Course::find($id);
+            $course->title = $request->title;
+            $course->price = $request->price;
+            $course->master_name = $request->master_name;
+            $course->master_job = $request->master_job;
+            if ($request->link !== null) {
+                $course->link = $request->link;
+            }
+            $course->introduction_v_link = $request->introduction_v_link;
+            if ($request->off > 0) {
+                $course->off = $request->off;
+            }
+            $course->type = $request->type;
+            $course->mode = $request->mode;
+            $course->meta_key_words = $request->meta_key_words;
+            $course->meta_descriptions = $request->meta_descriptions;
+            $course->introduction = $request->introduction;
+            $course->description = $request->description;
+            $course->licensable = $request->licensable;
+            $course->save();
+            return redirect()->route("admin.courses.all")->with("success", ".دوره شما با موفقیت ویرایش شد");
         }
-        $course->introduction_v_link = $request->introduction_v_link;
-        if ($request->off > 0) {
-            $course->off = $request->off;
+
+
+        // check if course is english
+        if ($request->lang == 1) {
+            $course = EnglishCourse::find($id);
+            $course->title = $request->title;
+            $course->price = $request->price;
+            $course->master_name = $request->master_name;
+            $course->master_job = $request->master_job;
+            if ($request->link !== null) {
+                $course->link = $request->link;
+            }
+            $course->introduction_v_link = $request->introduction_v_link;
+            if ($request->off > 0) {
+                $course->off = $request->off;
+            }
+            $course->type = $request->type;
+            $course->mode = $request->mode;
+            $course->meta_key_words = $request->meta_key_words;
+            $course->meta_descriptions = $request->meta_descriptions;
+            $course->introduction = $request->introduction;
+            $course->description = $request->description;
+            $course->licensable = $request->licensable;
+            $course->save();
+            return redirect()->route("admin.courses.all")->with("success", ".دوره شما با موفقیت ویرایش شد");
         }
-        $course->type = $request->type;
-        $course->mode = $request->mode;
-        $course->meta_key_words = $request->meta_key_words;
-        $course->meta_descriptions = $request->meta_descriptions;
-        $course->introduction = $request->introduction;
-        $course->description = $request->description;
-        $course->licensable = $request->licensable;
-        $course->save();
-        return redirect()->route("admin.courses.all")->with("success", ".دوره شما با موفقیت ویرایش شد");
     }
 
     /**
@@ -177,17 +261,32 @@ class AdminCourseController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-
-        $course = Course::find($id);
-        $path = pathinfo($course->images[0]->path)["dirname"];
-        foreach ($course->images as $image) {
-            File::delete($image->path);
+        // check if course in farsi
+        if ($request->lang == 0) {
+            $course = Course::find($id);
+            $path = pathinfo($course->images[0]->path)["dirname"];
+            foreach ($course->images as $image) {
+                File::delete($image->path);
+            }
+            rmdir($path);
+            $course->images()->delete();
+            $course->delete();
+            return redirect()->back()->with("success", ".دوره شما با موفقیت حذف شد");
         }
-        rmdir($path);
-        $course->images()->delete();
-        $course->delete();
-        return redirect()->back()->with("success", ".دوره شما با موفقیت حذف شد");
+
+        // check if course in english
+        if ($request->lang == 1) {
+            $course = EnglishCourse::find($id);
+            $path = pathinfo($course->images[0]->path)["dirname"];
+            foreach ($course->images as $image) {
+                File::delete($image->path);
+            }
+            rmdir($path);
+            $course->images()->delete();
+            $course->delete();
+            return redirect()->back()->with("success", ".دوره شما با موفقیت حذف شد");
+        }
     }
 }
